Adapt Flutter APP to the monorepo and updated doc (#59)
* Update doc and deploy reference

* Update doc and fix .env

* Minor bug fix

* Update README.md

* Fix github actions

* Reorganizing
