--- APP/.github/workflows/flutter.yml
@@ -1,30 +1,0 @@
-name: Flutter
-
-on:
-  push:
-    branches: [ "main" ]
-  pull_request:
-    branches: [ "main" ]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./ogree_app
-        
-    steps:
-      - uses: actions/checkout@v3
-      - uses: subosito/flutter-action@v2
-        with:
-          flutter-version: '3.7.6'
-          channel: 'stable'
-          
-      - name: Set up 
-        run: flutter pub get
-      
-      - name: Test
-        run: flutter test
-      
-      - name: Web build
-        run: flutter build web
